requirements.txt
cffi==1.15.1
chardet==5.0.0
charset-normalizer==3.0.0
cryptography==38.0.3
cssselect==1.2.0
distro==1.8.0
lxml==4.9.1
numpy==1.23.4
pandas==1.5.1
pdfminer.six==20221105
pdfquery==0.4.3
pycparser==2.21
pyquery==1.4.3
python-dateutil==2.8.2
pytz==2022.6
roman==3.3
six==1.16.0
tabula-py==2.5.1
